requirements.txt
appdirs==1.4.4
asgiref==3.3.4
distlib==0.3.2
Django==3.1.7
django-crispy-forms==1.11.1
django-filter==2.4.0
et-xmlfile==1.1.0
filelock==3.0.12
gunicorn==20.1.0
openpyxl==3.0.7
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
virtualenv==20.4.7
whitenoise==5.3.0
